refactor: controll -> control
